pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "google-calendar-mcp"
version = "0.1.0"
description = "MCP server exposing Google Calendar as tools for Claude"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
google-calendar-mcp  = "google_calendar_mcp.server:main"
google-calendar-auth = "google_calendar_mcp.server:auth_main"

[tool.hatch.build.targets.wheel]
packages = ["src/google_calendar_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
